refactor(trusty-mpm): split doctor_output_style tests into their own file
The 500-SLOC production file cap flagged doctor_output_style.rs at 552 SLOC
after issue #3453 added the precedence-chain layers and the legacy_ids probe.
Move the inline `mod tests` body to doctor_output_style_tests.rs (test-file
budget: 1500 SLOC), included via #[path], matching the existing
doctor.rs/doctor_tests.rs split. No behavior change.
